notification/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.2.0
sqlalchemy[asyncio]==2.0.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-multipart==0.0.9
ruff==0.3.0
pika==1.3.2
python-telegram-bot==20.7
